--- src/routes/graphql/dataloaders/all.ts
@@ -1,1 +1,0 @@
-export * from './user.dataloader.js';
