+++ src/routes/graphql/member-type/member-type.model.ts
