--- src/routes/graphql/models/all.ts
@@ -1,2 +1,0 @@
-export * from './user.model.js';
-export * from './member-type.model.js';
