--- src/routes/graphql/queries/all.ts
@@ -1,2 +1,0 @@
-export * from './user.query.js';
-export * from './member-type.query.js';
