+++ src/routes/graphql/types/all.ts
@@ -1,3 +1,2 @@
 export * from './uuid.js';
-export * from './user.type.js';
 export * from './context.type.js';
