+++ src/routes/graphql/types/context.type.ts
@@ -1,7 +1,7 @@
 import { PrismaClient } from '@prisma/client';
 import DataLoader from 'dataloader';
-import { UserModel } from '../models/user.model.js';
-import { MemberTypeModel } from '../models/member-type.model.js';
+import { UserModel } from '../user/user.model.js';
+import { MemberTypeModel } from '../member-type/member-type.model.js';
 
 export type ContextType = {
   prismaClient: PrismaClient;
