+++ src/routes/graphql/user/user.model.ts
